Update README
Swift Package Manager requires an explicit package name. It happens since the name deduced from the URL (`swift-tagged`) defers from the name declared in the `Package.swift` file for the version `0.5.0` which is `Tagged`.

If the target dependency is declared as `.product(name: "Tagged", package: "Tagged")` and there is no explicit package name for package dependencies, the following error occurs: "unknown package 'Tagged' in dependencies of target '<target-name>'; valid packages are: 'swift-tagged'".

If the target dependency is declared as `.product(name: "Tagged", package: "swift-tagged")` and there is no explicit package name for package dependencies, the following error occurs: "product dependency 'Tagged' in package 'swift-tagged' not found".